// frontend/src/lib/sdk/models/CustomerContactResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CustomerContactResponse = {
    id: number;
    customer_id: number;
    contact_person: string;
    phone: string;
    email: (string | null);
    is_primary: boolean;
    created_at: string;
    updated_at: string;
};