// frontend/src/lib/sdk/models/CustomerAddressResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CustomerAddressResponse = {
    id: number;
    customer_id: number;
    type: string;
    address: string;
    is_pickup_default: boolean;
    is_delivery_default: boolean;
    created_at: string;
    updated_at: string;
};