// frontend/src/lib/sdk/models/CustomerContactUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CustomerContactUpdate = {
    contact_person?: (string | null);
    /**
     * Phone number in E.164 format
     */
    phone?: (string | null);
    email?: (string | null);
    is_primary?: (boolean | null);
};