// frontend/src/lib/sdk/models/CustomerAddressUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CustomerAddressUpdate = {
    type?: (string | null);
    address?: (string | null);
    is_pickup_default?: (boolean | null);
    is_delivery_default?: (boolean | null);
};